requirements.txt
awscrt==0.30.0
awsiotsdk==1.27.0
